refactor: Answers 패키지 위치 수정, delete 메소드명 수정
