+++ src/main/java/qna/domain/answer/Answer.java
@@ -53,7 +53,7 @@ public class Answer extends BaseTimeEntity {
         return new Answer(question, writer, contents);
     }
 
-    public void changeDeleteState(User loginUser) throws CannotDeleteException {
+    public void delete(User loginUser) throws CannotDeleteException {
         if (!isOwner(loginUser)) {
             throw new CannotDeleteException("다른 사람이 쓴 답변이 있어 삭제할 수 없습니다.");
         }
